Implement hotel offer deals API and it's representational beans

+++ src/main/java/com/expedia/beans/HotelInfo.java
@@ -21,7 +21,7 @@ public class HotelInfo {
             imageUrl,
             starRating;
 
-    private float latitude,
+    private double latitude,
             longitude,
             guestReviewRating;
 
@@ -126,27 +126,27 @@ public class HotelInfo {
         this.starRating = starRating;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getGuestReviewRating() {
+    public double getGuestReviewRating() {
         return guestReviewRating;
     }
 
-    public void setGuestReviewRating(float guestReviewRating) {
+    public void setGuestReviewRating(double guestReviewRating) {
         this.guestReviewRating = guestReviewRating;
     }
